+++ Libft/ft_strlcat.c
@@ -15,16 +15,17 @@
 
 #include "libft.h"
 
+
 // #include <stdio.h>
 
 size_t	ft_strlen(const char *str);
 
 size_t	ft_strlcat(char *dst, const char *src, size_t n)
 {
-	int		i;
-	size_t	dst_len;
-	size_t	src_len;
-	size_t	remaining;
+	size_t i;
+	size_t dst_len;
+	size_t src_len;
+	size_t remaining;
 
 	i = 0;
 	dst_len = ft_strlen(dst);
